Rename `channe_manager.rs` to `channel_manager.rs`
... as this seems to be an obvious typo.
