+++ lampod/src/ln/mod.rs
@@ -1,5 +1,5 @@
 //! Lampo Channel Manager
-mod channe_manager;
+mod channel_manager;
 mod inventory_manager;
 mod offchain_manager;
 mod peer_manager;
@@ -7,7 +7,7 @@ mod peer_manager;
 pub mod events;
 pub mod peer_event;
 
-pub use channe_manager::LampoChannelManager;
+pub use channel_manager::LampoChannelManager;
 pub use inventory_manager::LampoInventoryManager;
 pub use offchain_manager::OffchainManager;
 pub use peer_manager::LampoPeerManager;
